+++ src/metric.ts
@@ -106,12 +106,8 @@ export function createMetrics(batches: Batch[]): FieldMetric[] {
     // Same source image, so same source image features. No need to compare.
     if (field.id === FieldId.WIDTH || field.id === FieldId.HEIGHT) continue;
     if (field.id === FieldId.FRAME_COUNT) continue;
-    if (field.id === FieldId.MEGAPIXELS) continue;
     // Encoder settings should not be compared.
-    if (field.id === FieldId.EFFORT || field.id === FieldId.QUALITY) continue;
-    // If bpp values are available, encoded sizes probably are too.
-    // Skip the former which brings nothing as a metric over the latter.
-    if (field.id === FieldId.ENCODED_BITS_PER_PIXEL) continue;
+    if (field.id === FieldId.EFFORT) continue;
 
     metrics.push(new FieldMetric(fieldIndices));
   }